backend/server/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
pandas>=2.2.0
reportlab>=4.2.0
gunicorn>=21.0.0
matplotlib>=3.8.0
numpy>=1.26.0
pillow>=10.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
